Move reflection marking helpers out of scanner
